# internal/docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact: {}
  description: API with Clean Architecture
  title: Indagio API
  version: "1.0"
paths:
...
  /projects/{id}/instruments:
    get:
      description: List instrument templates in a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List instruments
      tags:
      - Instruments
    post:
      consumes:
      - application/json
      description: Create an instrument template in a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Instrument payload
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create instrument
      tags:
      - Instruments
  /projects/{id}/instruments/{instrumentID}:
    patch:
      consumes:
      - application/json
      description: Update instrument name, config, or status
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Instrument ID
        in: path
        name: instrumentID
        required: true
        type: string
      - description: Instrument patch payload
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update instrument
      tags:
      - Instruments
...
  /projects/{id}/questionnaires:
    get:
      description: List questionnaire versions in a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List questionnaires
      tags:
      - Questionnaires
    post:
      consumes:
      - application/json
      description: Create a questionnaire schema for a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Questionnaire payload
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create questionnaire
      tags:
      - Questionnaires
  /projects/{id}/questionnaires/{questionnaireID}/publish:
    post:
      description: Publish a questionnaire version for active use
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Questionnaire ID
        in: path
        name: questionnaireID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Publish questionnaire
      tags:
      - Questionnaires
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"